// libs/features/contributor/src/lib/shared/enums/contribution-category.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum ContributionCategory {
  ANNUITY = 'Annuity',
  OH = 'Occupational Hazard',
  UI = 'Unemployment Insurance',
  OH_ANNUITY = 'Annuity and Occupational Hazard',
  PPA = "PPA Annuity",
  PENSION_REFORM_ANNUITY = 'Pension Reform Annuity'
}